just checking the checkin...
git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55100 6189c7af-7406-0410-a2ce-f22f51069c23

----
An additional 9 commit messages were squashed into this commit:
--
adding docking metrics

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55098 6189c7af-7406-0410-a2ce-f22f51069c23
--
removing broker- source-files that are no longer needed.

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55096 6189c7af-7406-0410-a2ce-f22f51069c23
--
refactoring of topology-broker
changes to noesy-assignment module

cosmetic changes in the following integration tests
FAIL broker_membrane
FAIL metalloprotein_broker
FAIL fold_and_dock
FAIL ThreadingInputter
FAIL fragment_picker
FAIL broker

expected changes in the following integration tests
FAIL noe_assignment

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55094 6189c7af-7406-0410-a2ce-f22f51069c23
--
fix input-files for ThreadingInputter broker-test

cosmetic changes...

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55092 6189c7af-7406-0410-a2ce-f22f51069c23
--
change output in broker integration test to make changes more apparent.

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55091 6189c7af-7406-0410-a2ce-f22f51069c23
--
allowing run of single integration test without deleting the whole 'new' directory

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55090 6189c7af-7406-0410-a2ce-f22f51069c23
--
remove unused input directory

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55089 6189c7af-7406-0410-a2ce-f22f51069c23
--
applying ignore list also to databaes dumps, because they contain path-names to the database...

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55087 6189c7af-7406-0410-a2ce-f22f51069c23
--
improving ingore list to get rid of unwanted integration test changes based on different run-paths

git-svn-id: https://svn.rosettacommons.org/source/trunk/rosetta@55086 6189c7af-7406-0410-a2ce-f22f51069c23
--

Old repository SHA1: a4d47c8aa87d2322a80525cb21dd2e6c7d8dbf40

+++ tests/integration/integration.py
@@ -228,19 +228,24 @@ rm -r ref/; ./integration.py    # create reference results using only default se
     else: rename_to_ref = False
     outdir = "new"
 
-    if not options.compareonly:
-        # Remove everything in the current outdir, then re-create it empty
-        if path.isdir(outdir): shutil.rmtree(outdir)
-        os.mkdir(outdir)
-
     # Each test consists of a directory with a "command" file in it.
     if len(args) > 0:
         tests = args
     else:
         tests = [ d for d in os.listdir("tests") if not d.startswith(".") and path.isdir(path.join("tests", d)) ]
-
+  
     if not options.unordered:
         tests = order_tests(tests)
+    
+    if not options.compareonly:
+        if len(args)>0 and path.isdir(outdir): #we have individual tests... only remove single test directories
+            for test in tests:
+                testdir=outdir+'/'+test
+                if path.isdir(testdir): shutil.rmtree(testdir)
+        else:
+        # Remove everything in the current outdir, then re-create it empty
+            if path.isdir(outdir): shutil.rmtree(outdir)
+            os.mkdir(outdir)  
 
     runtimes={}
     if not options.compareonly:
